divisao criaçao data
