+++ template-labook/package.json
@@ -12,11 +12,11 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "cors": "^2.8.5",
     "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "knex": "^2.4.2",
     "mysql": "^2.18.1",
-    "cors": "^2.8.5",
     "typescript": "^4.9.4"
   },
   "devDependencies": {
