+++ template-labook/src/data/UserDatabase.ts
@@ -1,0 +1,10 @@
+import { BaseDatabase } from "./BaseDatabase";
+
+export class UserDatabase extends BaseDatabase {
+
+    
+
+
+
+
+}
