+++ template-labook/src/index.ts
@@ -1,4 +1,4 @@
-import express, { Express, Request, Response } from "express"
+import express, { Request, Response } from "express"
 import cors from "cors"
 import knex from "knex"
 import dotenv from "dotenv"
@@ -6,50 +6,12 @@ import Knex from "knex"
 
 /**************************** CONFIG ******************************/
 
-dotenv.config()
-
-export const connection= knex({
-   client: "mysql",
-   connection: {
-      host: process.env.DB_HOST,
-      user: process.env.DB_USER,
-      password: process.env.DB_PASS,
-      database: process.env.DB_NAME,
-      port: 3306,
-      multipleStatements: true
-   }
-})
-
 const app: Express = express()
 app.use(express.json())
 app.use(cors())
 
 /**************************** TYPES ******************************/
 
-type authenticationData = {
-   id: string
-}
-
-type user = {
-   id: string,
-   name: string,
-   email: string,
-   password: string
-}
-
-enum POST_TYPES {
-   NORMAL = "normal",
-   EVENT = "event"
-}
-
-type post = {
-   id: string,
-   photo: string,
-   description: string,
-   type: POST_TYPES,
-   createdAt: Date,
-   authorId: string
-}
 
 
 /**************************** ENDPOINTS ******************************/
